Ex Nihilo (by Ghennkins)
**Ghennkins's textures ported from FTB thread (page 53)**

"-Most the textures are modifications of existing JohnSmith textures.
-I included the already available items (crooks, hammers) just for
completion sake.
-The Crucibles were a pita to texture. I wanted to do something a little
fancy with them but the mod author made the texture sheet in in a very
unfriendly-to-texture way. I included a template if anyone is interested
in making their own crucible textures. On mine, there is 1 corner that
doesn't look right but to fix it I'd have to redo the whole thing and
\laaaazzyyy
-The Stone Barrel bottom textures are a bit messed up. Should be dark
instead of light. Maybe RP that your char sucks at pottery? =)
-I should have made the gravel ore item (looks like coal) brighter and a
different color so it didn't look like coal with stripes but again
/laaaaaaazzzyyyy n_n
-I saw video of Sphax textures for AR2 magic water with like ghosts in
it that I really liked the idea of. So I tried to do something similar
for Witch Water. If you hate it, it's just one of JS's water textures
colored purple so shouldn't be too difficult replacing. I wish I could
have done something for the flowing water too that would be fitting."